BSP refactoring
