+++ grails-app/conf/spring/resources.groovy
@@ -1,7 +1,6 @@
 package spring
 
 import org.broadinstitute.orsp.config.AppInfoConfiguration
-import org.broadinstitute.orsp.config.BspConfiguration
 import org.broadinstitute.orsp.config.ConsentConfiguration
 import org.broadinstitute.orsp.config.BQConfiguration
 import org.broadinstitute.orsp.config.DataBioConfiguration
@@ -13,7 +12,6 @@ beans = {
 
     // Grails/Spring doesn't seem to be able to autowire these configuration beans otherwise
     appInfoConfiguration(AppInfoConfiguration) {}
-    bspConfiguration(BspConfiguration) {}
     consentConfiguration(ConsentConfiguration) {}
     bqConfiguration(BQConfiguration) {}
     dataBioConfiguration(DataBioConfiguration) {}
