--- src/main/groovy/org/broadinstitute/orsp/config/BspConfiguration.groovy
@@ -1,18 +1,0 @@
-package org.broadinstitute.orsp.config
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-@ConfigurationProperties(prefix = "bsp")
-class BspConfiguration {
-    Boolean sync
-    BspServiceConfiguration service
-}
-
-class BspServiceConfiguration {
-    String username
-    String password
-    String allSampleCollectionsUrl
-    String statusUrl
-}
